Move ert.shared.dark_storage to ert.dark_storage
