+++ .mypy-strict.ini
@@ -5,6 +5,7 @@ exclude = (?x)(
     src/ert/logging
     | src/ert/ert3/examples
     | src/ert/shared
+    | src/ert/dark_storage
     | src/ert/_c_wrappers
   )
