+++ src/ert/dark_storage/endpoints/compute/misfits.py
@@ -8,8 +8,8 @@ from ert_storage.compute import calculate_misfits_from_pandas
 from fastapi import APIRouter, Depends, status
 from fastapi.responses import Response
 
-from ert.shared.dark_storage.common import data_for_key, observations_for_obs_keys
-from ert.shared.dark_storage.enkf import LibresFacade, get_name, get_res
+from ert.dark_storage.common import data_for_key, observations_for_obs_keys
+from ert.dark_storage.enkf import LibresFacade, get_name, get_res
 
 router = APIRouter(tags=["misfits"])
