+++ src/ert/dark_storage/endpoints/observations.py
@@ -4,7 +4,7 @@ from uuid import UUID, uuid4
 from ert_storage import json_schema as js
 from fastapi import APIRouter, Body, Depends
 
-from ert.shared.dark_storage.enkf import LibresFacade, get_res
+from ert.dark_storage.enkf import LibresFacade, get_res
 from ert.shared.storage.extraction import create_observations
 
 router = APIRouter(tags=["ensemble"])
