+++ src/ert/dark_storage/endpoints/server.py
@@ -3,7 +3,7 @@ from typing import Any, Mapping
 
 from fastapi import APIRouter, Depends
 
-from ert.shared.dark_storage.enkf import LibresFacade, get_res
+from ert.dark_storage.enkf import LibresFacade, get_res
 
 router = APIRouter(tags=["info"])
