+++ tests/performance_tests/performance_utils.py
@@ -12,7 +12,7 @@ from ecl.summary import EclSum
 from jinja2 import Environment, FileSystemLoader
 from numpy import array
 
-from ert.shared.dark_storage import enkf
+from ert.dark_storage import enkf
 
 
 def write_summary_spec(file, keywords):
@@ -140,7 +140,7 @@ def dark_storage_app(monkeypatch):
     monkeypatch.setenv("ERT_STORAGE_NO_TOKEN", "yup")
     monkeypatch.setenv("ERT_STORAGE_RES_CONFIG", "poly.ert")
     monkeypatch.setenv("ERT_STORAGE_DATABASE_URL", "sqlite://")
-    from ert.shared.dark_storage.app import app
+    from ert.dark_storage.app import app
 
     yield app
     reset_enkf()
