+++ tests/performance_tests/test_dark_storage_performance.py
@@ -5,8 +5,8 @@ import pytest
 
 from ert._c_wrappers.enkf import EnKFMain, ResConfig
 from ert.async_utils import run_in_loop
+from ert.dark_storage.endpoints import ensembles, experiments, records, responses
 from ert.libres_facade import LibresFacade
-from ert.shared.dark_storage.endpoints import ensembles, experiments, records, responses
 
 
 def get_single_record_csv(ert, ensemble_id1, keyword, poly_ran):
